+++ src/components/index.ts
@@ -1,3 +1,3 @@
 export { EnterMeetingRoom } from "./modal/enter-meeting_room/enter-meeting_room";
-export { HomeSideBar } from "./sideBar/homeSideBar";
-export { MeetingRoomSideBar } from "./sideBar/meetingRoomSideBar";
+export { HomeSideBar } from "./sideBar/HomeSideBar";
+export { MeetingRoomSideBar } from "./sideBar/MeetingRoomSideBar";
